secboot: generate a efi load sequence for recover
During install the environment is more controlled and we know that
the recover mode kernel is the same that will be extracted and used
in run mode, but to match the profile used for resealing we should
also create a recover mode load sequence. Having matching profiles
could make it easier to verify if the policy has changed before
resealing.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>